feat(otlp-grpc-exporter-otlp): clean up
